+++ src/ast.rs
@@ -103,6 +103,16 @@ pub enum Statement {
         span: Option<SourceSpan>,
     },
 
+    /// Module metadata declaration: `module.metadata = (...)`
+    ModuleMetadata {
+        version: Option<String>,
+        description: Option<String>,
+        author: Option<String>,
+        license: Option<String>,
+        doc_comments: Option<Vec<String>>,
+        span: Option<SourceSpan>,
+    },
+
     /// Module interface declaration: `module.interface = (...)`
     ModuleInterface {
         inputs: HashMap<String, ModuleInput>,
@@ -118,11 +128,14 @@ pub enum Statement {
         span: Option<SourceSpan>,
     },
 
-    /// Module instantiation: `module.<type>.<instance> = (source = "...", ...)`
+    /// Module instantiation: `module.<type>.<instance> = (source = "...", count = N, for_each = [...], ...)`
     ModuleInstance {
         module_type: String,
         instance_name: String,
         source: String,
+        when: Option<Expression>, // Named 'when' for brevity, uses 'condition' in syntax
+        count: Option<Expression>, // Create N instances
+        for_each: Option<Expression>, // Create instances for each element
         inputs: HashMap<String, Expression>,
         doc_comments: Option<Vec<String>>,
         span: Option<SourceSpan>,
@@ -362,6 +375,7 @@ impl Statement {
             Statement::Import { span, .. } => span.as_ref(),
             Statement::ForLoop { span, .. } => span.as_ref(),
             Statement::Expression { span, .. } => span.as_ref(),
+            Statement::ModuleMetadata { span, .. } => span.as_ref(),
             Statement::ModuleInterface { span, .. } => span.as_ref(),
             Statement::ModuleOutputs { span, .. } => span.as_ref(),
             Statement::ModuleInstance { span, .. } => span.as_ref(),